Moving to conf system - CoordinateGridMixin 03
